requirements.txt
pyyaml
python-magic-bin
groq
python-docx
openpyxl
PyPDF2
pandas